made navigation more nested, consisted, and concise. Even renamed some pages for clarity
