+++ client/src/App.tsx
@@ -4,7 +4,7 @@ import '@/SCSS/PageAndHome/Page.scss';
 import '@/SCSS/NavHeaderFooterError/Navigation.scss';
 import '@/SCSS/NavHeaderFooterError/Error.scss';
 import '@/SCSS/NavHeaderFooterError/Footer.scss'
-import Footer from '@/components/Footer'
+import Footer from '@/components/Shared/Footer'
 
 export default function App() {
