+++ client/src/components/Brands/Map/MapDisplay.tsx
