+++ client/src/components/CarInformation/CarDetails/CarImage.tsx
@@ -1,4 +1,4 @@
-import { Car } from "@/CarDetails/CarInterfaces";
+import { Car } from "@/components/CarInformation/CarDetails/CarInterfaces";
 
 // Define backend URL for image fetching (same as in ImageCarousel.tsx)
 const backendImageUrl = import.meta.env.VITE_PUBLIC_BASE_URL ?? "http://localhost:3001";
