+++ client/src/components/CarInformation/StarRank.tsx
