+++ client/src/components/HomePage/GameInfo.tsx
