+++ client/src/components/Shared/Footer.tsx
