--- client/src/pages/Manufacturers.tsx
@@ -1,24 +1,0 @@
-import Header from "@/components/Header";
-import JumpList from "@/Manufacturers/AlphabetJumpList";
-import PageTab from '@/components/PageTab';
-import '@/CSS/Manufacturer.css'
-
-export default function Manufacturers() {
-
-    return (
-
-        <>
-
-            <div>
-
-                <PageTab title="Brands">
-                    <Header text="Brands" />
-                    <JumpList />
-                </PageTab>
-
-            </div>
-
-        </>
-
-    )
-}
