+++ client/src/routes/index.tsx
@@ -1,24 +1,24 @@
 import { createBrowserRouter } from "react-router-dom";
 import App from "@/App";
-import Home from '@/pages/Home'
-import ErrorPage from "@/pages/ErrorPage";
+import Home from '@/pages/MainPages/Home/Home'
+import ErrorPage from "@/pages/Error/ErrorPage";
 // import Manufacturers from "@/pages/Manufacturers";
 
 // ================================================================================
 //                          Manufacturers Page Start
 // ================================================================================
 
-import Brands from "@/pages/Manufacturers/Brands";
-import BrandInfo from "@/pages/Manufacturers/BrandInfo";
+import Brands from "@/pages/MainPages/Brands/Brands";
+import BrandInfo from "@/pages/MainPages/Brands/BrandInfo";
 
 // ================================================================================
 //                          Manufacturers Page End
 // ================================================================================
 
-import CarsByClass from "@/pages/CarInfo/CarsByClass";
-import CarDetail from "@/pages/CarInfo/CarDetail";
-import GarageLevels from "@/pages/GarageLevels";
-import LegendStorePrices from "@/pages/LegendStore";
+import Cars from "@/pages/MainPages/CarInfo/Cars";
+import CarDetails from "@/pages/MainPages/CarInfo/CarDetails";
+import GarageLevels from "@/pages/MainPages/GarageLevels/GarageLevels";
+import LegendStorePrices from "@/pages/MainPages/LegendStore/LegendStore";
 
 export const router = createBrowserRouter([
     {
@@ -31,20 +31,20 @@ export const router = createBrowserRouter([
           element: <Home />,
         },
         {
-          path: "/manufacturers/manufacturersmap", // temporary until you rename later
+          path: "/brands", // temporary until you rename later
           element: <Brands />,
         },
         {
-          path: "/manufacturers/:slug", // 🔥 dynamic route!
+          path: "/brands/:slug", // 🔥 dynamic route!
           element: <BrandInfo />, // ✅ Correct now - BrandInfo is the full brand page
         },
         {
-          path: "/carsbyclass",
-          element: <CarsByClass />,
+          path: "/cars",
+          element: <Cars />,
         },
         {
           path: "/cars/:id",
-          element: <CarDetail />,
+          element: <CarDetails />,
         },
         {
           path: "/garagelevels",
